Fix position calculation for highlight menu
Previously, highlight menu was positioned using position: absolute
and using the getBoundingClientRect() of the last selected range
but this caused highlight menu to not be positioned at the tail of
the selection when the selection spanned multiple lines of a single
text node (causing the menu to be at the far right end of the
selection). This fix instead inserts an inline-flex div at the end
of the range and calls normalize() on the parent element of the text
node which it has split upon being removed (to maintain structural
integrity).