Fix for highlight altering para structure sometimes
Previously, highlight menu was directly inserted at either
the start or the end of the selection, thus splitting the
selection apart from its neighboring content. This caused
visible shifts in content structure (e.g. a character being
moved up to the previous line to accomodate the menu).
To fix this, the highlight menu is now appended to the body
instead and positioned absolutely by finding the coordinates
of a temporary span placed at the end/start of the selection.