Exclude whitespaces at start and end of selection
This fixes the incorrection positioning of the highlight
menu when the user extends the selection to the very end
of the previous line (in which case the menu floats above
the previous line) or to the very start of the next line
(in which case the menu floats below the next line). This
was caused by the user selecting an invisible, non-highlightable
whitespace character in the preceding line end or the
proceeding line start.