Refactor highlight menu code
Previously, highlight menu was created every time the selection changed
but this was unnecessary work. The code has been refactored so that
highlight menu is initialized and hidden at the start, and only shown
and positioned appropriately once the user selects something.

The internals of the menu has also been changed to use shadow DOM for
tighter encapsulation and separation from the main document.

This commit also fixes some finer-grained calculation errors in the left
and top offset calculation for the menu.