Fix element not normalizing after menu pos. shift
In previous releases, there was an issue whereby the element
affected by the break in the DOM structure caused by the insertion
of the zero-width span (as a temporary marker for positioning the menu)
was not normalized properly when the menu's vertical orientation had
to be shifted as a result of the menu not being visible in the window.

Since range.normalize() was only called once when the menu was closed,
but there were two breaks caused by the span in this case, the original
DOM structure was not being preserved.