Persist user highlights
This makes use of chrome's storage API. Pretty hacky right now, need to
clean up later.

Here are 2 particular implementation details to take note of:
- It's important that we do normalize() before serializing selection
range for persistence (in app.js)
- When generating range from highlightInfo, we need to normalize() the
parent element of either the startContainer or the endContainer, or both
(in case the highlightMenu position was swapped to keep it in window
view). This ensures that the DOM structure is the same as when we
manually highlight some text, otherwise since there is no menu generated
for highlighting directly from our persisted data, there will be no
normalization taking place which would normally get rid of the
whitespaces automatically added by HTML between our custom
'highlight-snippet' element and other elements/nodes.