Refactor codebase
Refactor code which deals with the creation of SelectionRange,
particularly with regards to how the appropriate range is determined.
For example, if the startContainer of the selected range is an element,
it finds the next closest text node in the DOM and sets the range start
to the start of that node (same goes for endContainer).

getHighlightRanges() now also takes care of excluding the whitespaces
at the start and end of a given text node within the selected range.